Refine prefix analysis.
The prefix analysis would lemmatize 'zustanden stehen' to 'stehen'
rather than 'zu#stehen'. The longest prefix search found 'zustande' as
a prefix of 'zustanden'. This prefix was then rejected because the
remaining form would be the character 'n'.

We now generate all possible prefix sequences for the form, retain
those that qualify to various heuristics and then take the prefix
sequence that leaves the shortest form and is the longest prefix
(e.g. preferring dabei# over da#bei#).