Improve marking of prefixes in VVIZU verbs.
Before, we used the longest set of prefixes and removed the last
prefix when it was 'zu'. However, this led to incorrect analyses of
e.g. 'hinzubewegen' as 'hinzu#bewegen' rather than 'hin#bewegen'.

The handling is now generalized: we exclude any prefix sequence
where:

1. The last prefix ends in '-zu'.
2. The remaining form does not start with 'zu-'.

The second condition is important to correctly lemmatize inflections
such as 'zuzuspitzen' into 'zu#spitzen'.