Detect one case of auxiliary/modal conjunctions.
When there is a coordination of two auxiliary or modal verbs, only one
of the two verbs will have a dependent with the AUX relation.

These changes detect whether an auxiliary/modal is an a conjunction
with another auxiliary/modal that has an auxiliary dependent. However,
we only add the tag for verbs that do not have any dependents (except
punctuation), because two VA.* or VM.* verbs can also occur in
conjunctions where the dependent are not shared.