RemoveTruncMarker: Use UPOS to decide to lowercase truncations
This RemoveTruncMarker transformation did not fire for the CoNLL-U
TüBa-D/Z, since it expected a percentage sign in the lemma. However,
the part-of-speech is not encoded in the lemma in the CoNLL-U
version. So, we always use the form as the lemma for truncations and
rely on the UPOS tag to decide whether to lowercase the lemma.