Cut down the separable verb particle list
The previous list was generated by finding all lemmas of the form

particle#lemma

in TüBa-D/Z. This gives some infrequent cases that are not particles
but compounds. Then there was noise such as 're', which resulted in
partitioning verbs that start with 're-'.

The new list is constructed by looking at particles that are actually
separated and occur at least three times (to exclude some annotation
errors).